Split account-group helpers into pure + server modules (FINLYNQ-179)
The cycle-1 commit (82f75e9) failed the Turbopack production build with
'Module not found: Can't resolve dns': src/lib/accounts/groups.ts
co-located the pure helpers (ACCOUNT_GROUP_DEFAULTS, parseGroupOrder,
groupSuggestions, orderGroups, dedupeGroups) WITH the server-side DB
functions (settings get/set + the owner-scoped renameAccountGroup). The
"use client" accounts page and its child components import the pure
helpers from that module, which transitively pulled the db layer -> pg
into the browser bundle, and pg's connection-parameters needs Node's dns.

Move the DB/settings/rename functions into a new server-only module
src/lib/accounts/groups-server.ts (imported only by the two API routes)
and keep groups.ts pure + client-safe. normalizeGroupName is now
exported from the pure module so the server module reuses the same
rename guard. No behavior change; the wire contracts and the pure cores
are byte-identical.

Verified with the FULL build this time: npm run build exit 0, npm run
lint 0 errors, npm run audit:invariants 15/15 PASS, and the 12
account-groups unit tests pass.
